requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.9.2
bcrypt>=4.2.0
python-multipart>=0.0.12
jinja2>=3.1.4
openpyxl>=3.1.5
python-dotenv>=1.0.1
itsdangerous>=2.2.0
qrcode[pil]>=7.4.2
python-jose[cryptography]>=3.3.0
mcp>=1.0.0
httpx>=0.27.0
